+++ Source/Shaders/CentralBodyVS.glsl
@@ -6,7 +6,6 @@ uniform int u_mode;
 
 uniform vec3 u_center3D;
 uniform mat4 u_modifiedModelView;
-uniform mat4 u_modifiedModelViewProjection;
 uniform vec4 u_tileExtent;
 
 // Uniforms for 2D Mercator projection
@@ -30,7 +29,7 @@ float get2DYPositionFraction();
 
 vec4 getPosition3DMode(vec3 position3DWC)
 {
-    return u_modifiedModelViewProjection * vec4(position3D, 1.0);
+    return czm_projection * u_modifiedModelView * vec4(position3D, 1.0);
 }
 
 float get2DMercatorYPositionFraction()
@@ -70,7 +69,7 @@ vec4 getPosition2DMode(vec3 position3DWC)
 {
     float yPositionFraction = get2DYPositionFraction();
     vec4 rtcPosition2D = vec4(0.0, mix(u_tileExtent.st, u_tileExtent.pq, vec2(textureCoordinates.x, yPositionFraction)), 1.0);  
-    return u_modifiedModelViewProjection * rtcPosition2D;
+    return czm_projection * u_modifiedModelView * rtcPosition2D;
 }
 
 vec4 getPositionColumbusViewMode(vec3 position3DWC)