+++ CHANGES.md
@@ -22,6 +22,7 @@ Beta Releases
 * Added `enableTranslate`, `enableZoom`, and `enableRotate` properties to `Camera2DController` to selectively toggle camera behavior.  All values default to `true`.
 * Added `Camera2DController.setPositionCartographic` to simplify moving the camera programmatically when in 2D mode.
 * Added Matrix4.multiplyByTranslation, Matrix4.fromScale, and Matrix3.fromScale.
+* Improved near/far plane distances and eliminated z-fighting.
 
 ### b8 - 09/05/2012